+++ modtest/LuaFont.cpp
@@ -6,10 +6,10 @@ static int ValidateDrawStringBase(lua_State* L, void (Font::* fn)(DrawStringEntr
 	DrawStringEntry entry;
 	Font* font = lua::GetUserdata<Font*>(L, 1, lua::Metatables::FONT, "Font");
 	entry._text = luaL_checkstring(L, 2);
-	entry._x = luaL_checknumber(L, 3);
-	entry._y = luaL_checknumber(L, 4);
+	entry._x = (float)luaL_checknumber(L, 3);
+	entry._y = (float)luaL_checknumber(L, 4);
 	entry._color = *lua::GetUserdata<KColor*>(L, 5, lua::Metatables::KCOLOR, "KColor");
-	entry._boxWidth = luaL_optinteger(L, 6, -1);
+	entry._boxWidth = (int)luaL_optinteger(L, 6, -1);
 	entry._center = lua::luaL_optboolean(L, 7, false);
 
 	(font->*fn)(entry);
@@ -20,12 +20,12 @@ static int ValidateDrawStringScaledBase(lua_State* L, void (Font::* fn)(DrawStri
 	DrawStringScaledEntry entry;
 	Font* font = lua::GetUserdata<Font*>(L, 1, lua::Metatables::FONT, "Font");
 	entry._text = luaL_checkstring(L, 2);
-	entry._x = luaL_checknumber(L, 3);
-	entry._y = luaL_checknumber(L, 4);
-	entry._scaleX = luaL_checknumber(L, 5);
-	entry._scaleY = luaL_checknumber(L, 6);
+	entry._x = (float)luaL_checknumber(L, 3);
+	entry._y = (float)luaL_checknumber(L, 4);
+	entry._scaleX = (float)luaL_checknumber(L, 5);
+	entry._scaleY = (float)luaL_checknumber(L, 6);
 	entry._color = *lua::GetUserdata<KColor*>(L, 7, lua::Metatables::KCOLOR, "KColor");
-	entry._boxWidth = luaL_optinteger(L, 8, -1);
+	entry._boxWidth = (int)luaL_optinteger(L, 8, -1);
 	entry._center = lua::luaL_optboolean(L, 9, false);
 
 	(font->*fn)(entry);
